requirement.txt
absl-py==0.14.1
altair==4.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
base58==2.1.1
bleach==4.1.0
blinker==1.4
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
clang==5.0
click==7.1.2
colorama==0.4.4
comtypes==1.1.10
cryptography==36.0.1
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.3
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.6.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.3
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
jsonschema==4.3.0
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.6
numpy @ file:///C:/ci_310/numpy_and_numpy_base_1643798589088/work
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pathtools==0.1.2
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.18.1
psutil==5.9.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
Pygments==2.10.0
Pympler==0.9
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==301
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
roboflow==0.2.2
rsa==4.7.2
scikit-learn @ file:///C:/ci/scikit-learn_1642617276183/work
scipy @ file:///C:/ci/scipy_1641555170412/work
seaborn==0.11.2
selenium==3.141.0
Send2Trash==1.8.0
sentry-sdk==1.5.7
setproctitle==1.2.2
shortuuid==1.0.8
six==1.15.0
sklearn==0.0
smmap==5.0.0
streamlit==1.2.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thop==0.0.31.post2005241907
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
toml==0.10.2
toolz==0.11.2
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.63.0
traitlets==5.1.0
typing-extensions==3.7.4.3
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.6
validators==0.18.2
wandb==0.12.11
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wget==3.2
widgetsnbextension==3.5.2
wincertstore==0.2
wrapt==1.12.1
yaspin==2.1.0
zipp==3.4.1
